Some better behavior, DreamerV1 actor loss, still deteriorates fast
It starts to behave better, especially cleanrl style actor, but actions quickly go to an extreme. I suspect it's not a matter of an actor, something's wrong in the loop. Investigate everything. Lack of batch updates is not a problem. It deteriorates the same way all the time, no matter the actor architecture