more metrics, todo
added logging of more metrics to easier investigate the issues. For some reason it definitely doesnt learn still as it's supposed to. I even loaded a checkpoint with good trained world model, but didnt load actor, so it started fresh, and it quickly got worse. As if the critic evaluation was guiding it wrong. Hmmmm